Add files via upload
According to the description of the travel agency business processes, the following actors can be identified: client, tour operator, travel agent.			

	Further it is necessary to determine the options of using the system and give a description of each of them:
1) Searching for a tour - the information system provides a catalog of all available tours.
2) Viewing information about the tour - the catalog contains complete information about the tour (country, date of departure and arrival, number of days of stay, cost).
3) Booking a tour - you can book a tour that is presented in the catalog.
4) Leave wishes on the site - the client can describe the tour of interest.
5) Forming a tour - the tour operator books the hotel, tickets, and determines the cost of the tour.
6) Adding a tour - the tour operator can include a new tour in the catalog.
7) Sales forecasting - hypothesizing future sales.
8) Realizing the tour - the travel agent sells the tour to the client.
9) Contract finalization - preparation of necessary documents.

Lab 1			
Create a Precedent Diagram describing the subject area.





 


lab1.1_Task1  Construct a use case diagram for the system

Use case "Checkout":
Brief description
This use case describes the process of placing a customer order in the information system, received from the customer by phone.
Main event flow
An employee asks the customer for his/her personal information.
The employee checks if the information about the customer is available in the information system. If the information is not found, the alternative flow Entry of information about the client into the information system is executed.
Employee selects a customer record in the information system.
The employee enters the order information for the selected customer into the information system.
Alternative flow
Entering customer information into the information system
If during the execution of the main flow it is found that the customer has not been previously registered in the information system, the Employee registers the customer in the information system.
Prerequisites
This use case starts executing when the customer informs the Employee by phone that he/she wants to place an order.  
Postconditions
If the use case completes successfully, the order is placed.